+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- fix: purge stale KV cache on 403 in addition to 404
 ## v3.4.17 - 2026-03-26
 
 - fix: disabled campaigns to stay live indefinitely